pyproject.toml
[tool.poetry]
name = "slackapps-stech"
version = "0.1.0"
description = ""
authors = ["tamanobi <tamanobi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1.1"
requests = "^2.23.0"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
